Split app global event wiring
