+++ apps/desktop/src/renderer/app-types.ts
@@ -1,0 +1,1 @@
+export type AppView = 'home' | 'chat' | 'automations' | 'agents' | 'capabilities' | 'pulse'
